requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot==21.*